Projects: hdmi: correct hw_handoff contents
Replaced hdf file with one directly exported by Vivado. The hw_handoff
folder should now be manually specified when exporting hardware from
vivado. It also needs to be specified when launching sdk from Vivado,
and the workspace should be set to the sdk folder (not "local to
project"). This workflow will make releasing modifications to this
project more streamlined and automated, and ensure this hdf file always
gets updated.